requirements.txt
requests==2.28.1
beautifulsoup4==4.11.1
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.42.0
selenium==4.1.3
pandas==1.4.2